Keep trying
Description
Let's make the game iterative. It's time to make it resemble the classical Hangman a bit more: a player should guess letters in the word instead of typing the entire word at once. If the player guesses a letter, it should be uncovered in the word. For now, start with the defeat case and add 8 tries to guess a letter that appears in the word. When the player runs out of attempts, the game ends.

Later we will determine the winning conditions, but in this stage, let's see how well our player guesses the word on every attempt.

Objectives
Now your game should work the following way:

A player has exactly 8 tries and enters letters. If a player has more tries but he actually guessed the word, it doesn't mean anything.
If the letter doesn't occur in the word, the computer takes one try away, even if the user already inputted this letter before.
If the player doesn't have any more attempts, the game should end and the program should show a losing message. Otherwise, the player can continue to input letters.
Also, use our previous word list: 'python', 'java', 'kotlin', 'javascript' so that your program can be tested more reliably.